Group source snippet components in submodule
