+++ codespan-reporting/src/term/views/mod.rs
@@ -1,15 +1,11 @@
 mod diagnostic;
-mod gutter;
 mod header;
 mod locus;
 mod new_line;
-mod note;
 mod source_snippet;
 
 pub use self::diagnostic::*;
-pub use self::gutter::*;
 pub use self::header::*;
 pub use self::locus::*;
 pub use self::new_line::*;
-pub use self::note::*;
 pub use self::source_snippet::*;
